refactor: move hints logic to a new HintsController
